Added more POD explaining how things work. Added tests for tar file and collection metadata.
Switched from TODO to SKIP for tests not requiring h5repack (I was
unable to get Test::More's TODO feature to cope with tests returning
early).

Removed setting sample_id by file path for Fastq files because it
would not work on current GridION. This can be revisited when the
GridION is updated.