Ensure add_object retains current checksum behaviour
A change is being make in perl-irods-wrap >3.5.1 to make checksum
creation behaviour consistent. The new default for add_object is NOT
to create the checksum.

This change uses the optional $WTSI::NPG::iRODS::CALC_CHECKSUM
argument to retain current behaviour. (In fact, not setting this and
defaulting to the new behaviour does not impact on the tests
involved.)