Simplify the test environment
Remove the TEST_AUTHOR distinction and always run the iRODS
tests.

This is an iRODS package, so we always want to test using iRODS. iRODS
servers for testing are cheap to create from our container images. We
are the authors, so we always want to run all the tests.

Remove setting the iRODS 3.x environment variable