Switch from IPC::Run to IPC::Open3 when executing iquest
We observed that successive output lines from iquest were joined,
intermittently. Switching to Open3 which is much simpler and gives us
a chance of diagnosing the problem (assuming the problem is not in
IPC::Run itself)

Use Cwd in the tests to resolve the relative path of the test data. I
had to test this with iRODS clients in a Docker container (on MacOS,
which iRODS does not support) and having the path fully resolved
allows clients to find the files, which are bind-mounted.