Unified logic for setting target and alt_target.
alt_target iRODS metadata attribute value is now computed
in the same way as the target metadata attribute value.
To be consistent with prior practice, the alt_target value
is assigned only when it is true.

This resolves the issue of tag zero files, which are  produced
by the alternative process, having alt_target metadata attribute
value of 1.