Fix check for number of reads obtained via samtools
This check is intended to ensure that when n reads are requested that
n reads are obtained. The majority of BAM/CRAM files contain millions
of reads, so in most cases simply checking for n reads is correct. In
the cases where the file contains only fewer than n reads, the maximum
we can expect is the number of reads in the file, rather than n.